Loading custom roles on front-end.
